Create helper functions for Config
Mostly in preparation for the ICU4X PR; I want to avoid having to write the exact same code over there for reading Config info from a file.
